# apps/server/etapi.openapi.yaml
openapi: "3.0.3"
info:
    version: 1.0.0
    title: ETAPI
    description: External Trilium API
    contact:
        name: Trilium Notes Team
        email: contact@eliandoran.me
        url: https://triliumnotes.org
    license:
        name: GNU Affero General Public License v3.0 only
        url: https://www.gnu.org/licenses/agpl-3.0.en.html
servers:
    - url: http://localhost:37740/etapi
    - url: http://localhost:8080/etapi
security:
    - EtapiTokenAuth: []
    - EtapiBasicAuth: []
paths:
...
    /notes/{noteId}/undelete:
        parameters:
            - name: noteId
              in: path
              required: true
              schema:
                  $ref: "#/components/schemas/EntityId"
        post:
            description: Restore a deleted note. The note must be deleted and must have at least one undeleted parent.
            operationId: undeleteNote
            responses:
                "200":
                    description: note restored successfully
                    content:
                        application/json; charset=utf-8:
                            schema:
                                type: object
                                properties:
                                    success:
                                        type: boolean
                                        example: true
                default:
                    description: unexpected error
                    content:
                        application/json; charset=utf-8:
                            schema:
                                $ref: "#/components/schemas/Error"
...
components:
    securitySchemes:
        EtapiTokenAuth:
            type: apiKey
            in: header
            name: Authorization
        EtapiBasicAuth:
            type: http
            scheme: basic
            description: >
                Basic Auth where username is arbitrary string (e.g. "trilium", not checked),
                username is the ETAPI token.
                To emphasize, do not use Trilium password here (won't work), only the generated
                ETAPI token (from Options -> ETAPI)
    schemas:
...
        EntityId:
            type: string
            pattern: "[a-zA-Z0-9_]{4,32}"
            example: evnnmvHTCgIn
...
        Error:
            type: object
            required:
                - status
                - code
                - message
            properties:
                status:
                    type: integer
                    format: int32
                    description: HTTP status, identical to the one given in HTTP response
                    example: 400
                code:
                    type: string
                    description: stable string constant
                    example: NOTE_IS_PROTECTED
                message:
                    type: string
                    description: Human readable error, potentially with more details,
                    example: Note 'evnnmvHTCgIn' is protected and cannot be modified through ETAPI
...